# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.City:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Country:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provinces:
        items:
          $ref: '#/definitions/models.Province'
        type: array
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Currency:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      symbol:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Guest:
    properties:
      address:
        type: string
      cell_number:
        type: string
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      national_id:
        type: string
      passport_date_of_issue:
        type: string
      passport_expire_date:
        type: string
      passport_number:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Hotel:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      email_address:
        type: string
      extra_data:
        type: string
      fax_number:
        type: string
      hotel_grade:
        $ref: '#/definitions/models.HotelGrade'
      hotel_grade_id:
        type: integer
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: integer
      phone_number1:
        type: string
      phone_number2:
        type: string
      postal_code:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      thumbnails:
        items:
          $ref: '#/definitions/os.File'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      website:
        type: string
    type: object
  models.HotelGrade:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.HotelType:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      grades:
        items:
          $ref: '#/definitions/models.HotelGrade'
        type: array
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      payer:
        $ref: '#/definitions/models.Guest'
      payer_id:
        type: integer
      payment_date:
        type: string
      payment_type:
        type: integer
      reservation:
        $ref: '#/definitions/models.Reservation'
      reservation_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Province:
    properties:
      alias:
        type: string
      cities:
        items:
          $ref: '#/definitions/models.City'
        type: array
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.RateCode:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      currency:
        $ref: '#/definitions/models.Currency'
      currency_id:
        type: integer
      deleted_at:
        type: string
      hotel:
        $ref: '#/definitions/models.Hotel'
      hotel_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      rate_group:
        $ref: '#/definitions/models.RateGroup'
      rate_group_id:
        type: integer
      status:
        description: |-
          Guest       Guest     `json:"guest"  valid:"-"`
          GuestId     uint64    `json:"guest_id"  valid:"required"`
        type: boolean
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.RateGroup:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      hotel:
        $ref: '#/definitions/models.Hotel'
      hotel_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Reservation:
    properties:
      check_status:
        type: integer
      checkin_date:
        type: string
      checkout_date:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      guest_count:
        type: integer
      hotel:
        $ref: '#/definitions/models.Hotel'
      hotel_id:
        type: integer
      id:
        type: integer
      nights:
        type: number
      parent:
        $ref: '#/definitions/models.Reservation'
      parent_id:
        type: integer
      price:
        type: number
      rate_code:
        $ref: '#/definitions/models.RateCode'
      rate_code_id:
        type: integer
      request_key:
        type: string
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      sharers:
        items:
          $ref: '#/definitions/models.Sharer'
        type: array
      supervisor:
        $ref: '#/definitions/models.Guest'
      supervisor_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Room:
    properties:
      clean_status:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      max_beds:
        type: integer
      name:
        type: string
      room_type:
        $ref: '#/definitions/models.RoomType'
      room_type_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.RoomType:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      hotel:
        $ref: '#/definitions/models.Hotel'
      hotel_id:
        type: integer
      id:
        type: integer
      max_guest_count:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Sharer:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      guest:
        $ref: '#/definitions/models.Guest'
      guest_id:
        type: integer
      id:
        type: integer
      reservation_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      birth_date:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
  os.File:
    type: object
info:
  contact: {}
  description: Swagger documentation for reservation API .
  license:
    name: MIT
  title: Reservation API
  version: "1.0"
paths:
...
  /hotel-types:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HotelType'
            type: array
      summary: Delete HotelType
      tags:
      - HotelType
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HotelType'
            type: array
      summary: findAll HotelTypes
      tags:
      - HotelType
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: HotelType
        in: body
        name: HotelType
        required: true
        schema:
          $ref: '#/definitions/models.HotelType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelType'
      summary: create new HotelType
      tags:
      - HotelType
  /hotel-types/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelType'
      summary: find HotelType by id
      tags:
      - HotelType
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: HotelType
        in: body
        name: HotelType
        required: true
        schema:
          $ref: '#/definitions/models.HotelType'
      - description: HotelType
        in: body
        name: HotelType
        required: true
        schema:
          $ref: '#/definitions/models.HotelType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelType'
      summary: update HotelType
      tags:
      - HotelType
  /hotels:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hotel'
            type: array
      summary: Delete Hotel
      tags:
      - Hotel
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hotel'
            type: array
      summary: findAll Hotels
      tags:
      - Hotel
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: create new Hotel
      tags:
      - Hotel
  /hotels/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: find Hotel by id
      tags:
      - Hotel
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: update Hotel
      tags:
      - Hotel
  /payments:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      summary: Delete Payment
      tags:
      - Payment
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Payment
        in: body
        name: Payment
        required: true
        schema:
          $ref: '#/definitions/models.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      summary: create new Payment
      tags:
      - Payment
...
swagger: "2.0"